requirements.txt
anyio==3.6.1
asgiref==3.4.1
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
fastapi==0.78.0
h11==0.13.0
httptools==0.3.0
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.1.1
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0

# for database actions
SQLAlchemy~=1.4.37

# for password hashing
passlib~=1.7.4

# for making http request [http client]
requests~=2.27.1

# for xml operations [third party api integration dependecies]
xmltodict~=0.13.0

# for test case
pytest~=7.0.1